+++ tests/prog1.cpp
@@ -18,7 +18,7 @@ int main(void)
         
         for (int i = 0; i < image_width; i++) 
         {
-            auto pixel_color = color(double(i)/(image_width-1), double(j)/(image_height-1), 0);
+            auto pixel_color = color(float(i)/(image_width-1), float(j)/(image_height-1), 0);
 
             write_color(std::cout, pixel_color);
         }
